CB-680 Move Server IP and URL to Cluster level in response
